% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.jags.R
\name{default.priors}
\alias{default.priors}
\title{Sets default priors for JAGS model code}
\usage{
default.priors(fun = tloglin())
}
\arguments{
\item{fun}{An object of class \code{"timefun"} generated (see Details) using any of
\code{tloglin()}, \code{tpoly()}, \code{titp()}, \code{temax()}, \code{tfpoly()}, \code{tspline()} or \code{tuser()}}
}
\value{
A list, each element of which is a named JAGS snippet
corresponding to a prior in the MBNMA JAGS code.
}
\description{
This function creates JAGS code snippets for default MBNMA model priors.
}
\examples{
\donttest{
default.priors(fun=temax())

default.priors(fun=titp(p.expon=TRUE))
}

}
