% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{gen.parameters.to.save}
\alias{gen.parameters.to.save}
\title{Automatically generate parameters to save for a dose-response MBNMA model}
\usage{
gen.parameters.to.save(fun, model)
}
\arguments{
\item{fun}{is a character specifying a functional form to be assigned to the
time-course. Options are given in \code{details}.}

\item{model}{A JAGS model written as a character object}
}
\description{
Automatically generate parameters to save for a dose-response MBNMA model
}
