% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{genspline}
\alias{genspline}
\title{#' Converts JAGS data held in \code{class("mbnma")} object to a data.frame
#'
#' @return A data frame similar in format to the data frame in an \code{"mb.network"} object.
#' @noRd
jagstonetwork <- function(mbnma) {
checkmate::assertClass(mbnma, "mbnma")}
\usage{
genspline(x, spline = "rcs", knots = 3, degree = 2, max.time = max(x))
}
\arguments{
\item{x}{A numeric vector indicating all time points available in the dataset}

\item{spline}{Indicates the type of spline function. Can be either natural cubic spline (\code{"ns"}), restricted cubic
spline (\code{"rcs"}) or B-spline (\code{"bs"}).}

\item{degree}{a positive integer giving the degree of the polynomial from which the spline function is composed
(e.g. \code{degree=3} represents a cubic spline).}

\item{max.time}{A number indicating the maximum time between which to calculate the spline function.}
}
\description{
jagsdat <- mbnma$model$data()
}
\details{

}
