% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{get.latest.time}
\alias{get.latest.time}
\title{Create a dataset with the latest time point only}
\usage{
get.latest.time(network)
}
\arguments{
\item{network}{An object of class \code{"mb.network"}.}
}
\value{
A data frame in long format of responses at the latest time point in
each arm of each study.
}
\description{
Takes the latest time point from each arm in each study within an
\code{mb.network} object. Useful for network plots.
}
\examples{
# Using the alogliptin dataset
network <- mb.network(alog_pcfb)

# Generate a data frame with only the latest time point included in each study
get.latest.time(network)

}
