% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma-class.R
\name{get.relative}
\alias{get.relative}
\title{Calculates relative effects/mean differences at a particular time-point}
\usage{
get.relative(
  mbnma,
  time = max(mbnma$model.arg$jagsdata$time, na.rm = TRUE),
  treats = mbnma$network$treatments,
  classes = NULL
)
}
\arguments{
\item{time}{A numeric value for the time at which to estimate relative effects/mean differences.}

\item{treats}{A character vector of treatment names for which to calculate relative effects/mean differences.
Must be a subset of \code{mbnma$network$treatments}}

\item{classes}{A character vector of class names for which to calculate relative effects/mean differences from.
Must be a subset of \code{mbnma$network$classes}. Only works for class effect models.}
}
\value{
A list containing details of the results and an array of class \code{relative.array} that contains
MCMC results for the relative effects / mean differences between all treatments specified in \code{treats}
or all classes specified in \code{classes}
}
\description{
Uses mbnma time-course parameter estimates to calculate relative effects or mean
differences (depending on scale) between treatments or classes at a particular time-point.
Can be used to compare treatments evaluated in studies at different follow-up times.
}
