% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{nonparam.run}
\alias{nonparam.run}
\title{Run a non-parametric random walk model}
\usage{
nonparam.run(
  network,
  class = FALSE,
  method = "common",
  link = "identity",
  binvals = bintime(network$data.ab),
  ...
)
}
\arguments{
\item{network}{An object of class \code{"mb.network"}.}

\item{link}{Can take either \code{"identity"} (the default),
\code{"log"} (for modelling Ratios of Means \insertCite{friedrich2011}{MBNMAtime}) or
\code{"smd"} (for modelling Standardised Mean Differences - although this also corresponds to an identity link function).}

\item{binvals}{A numeric vector defining the boundaries of the time
bins.}

\item{...}{Options for plotting in \code{igraph}.}
}
\description{
Runs a non-parametric model that splits the data into different time-bins
and models a random walk process between them, following the method
of \insertCite{lu2007;textual}{MBNMAtime}.
}
\examples{
# Using the alogliptin dataset
network <- mb.network(alog_pcfb)

# Specify time bins to use for analysis
timebins <- c(0,6,12,20,40)

# Run a common effects non-parametric RW model
nonparam.run(network, binvals=timebins, method="common")

}
