% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{osteopain}
\alias{osteopain}
\title{Studies of pain relief medications for osteoarthritis}
\format{
A data frame with 417 rows and 7 variables:
\itemize{
\item \code{studyID} Study identifiers
\item \code{time} Numeric data indicating follow-up times
\item \code{y} Numeric data indicating the mean response for a given observation
\item \code{se} Numeric data indicating the standard error for a given observation
\item \code{treatment} Treatment identifiers as factors. Labels are shortened treatment names.
\item \code{arm} Arm identifiers coded for each study
\item \code{treatname} Character data giving the full names of each treatment
}
}
\source{
Pfizer Ltd.
}
\usage{
osteopain
}
\description{
A dataset containing results on the WOMAC pain scale (0-10) over time for studies investigating 29
treatments for pain relief in patients with osteoarthritis. Standard deviations have been imputed for
269 observations.
}
\details{
\code{osteopain} is a data frame in long format (one row per observation, arm and study),
with the variables \code{studyID}, \code{time}, \code{y}, \code{se}, \code{treatment}, \code{arm} and \code{treatname}.
}
\keyword{datasets}
