% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot.mb.predict}
\alias{plot.mb.predict}
\title{Plots predicted responses from a time-course MBNMA model}
\usage{
\method{plot}{mb.predict}(
  x,
  disp.obs = FALSE,
  overlay.ref = TRUE,
  overlay.nma = NULL,
  col = "blue",
  max.col.scale = NULL,
  treat.labs = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"mb.predict"} generated by
\code{predict("mbnma")}}

\item{disp.obs}{A boolean object to indicate whether to show shaded sections
of the plot for where there is observed data (\code{TRUE}) or not (\code{FALSE})}

\item{overlay.ref}{A boolean object indicating whether to overlay a line
showing the median network reference treatment response over time on the
plot (\code{TRUE}) or not (\code{FALSE}). The network reference treatment (treatment
\enumerate{
\item must be included in \code{predict} for this to display the network reference
treatment properly.
}}

\item{col}{A character indicating the colour to use for shading if \code{disp.obs}
is set to \code{TRUE}. Can be either \code{"blue"}, \code{"green"}, or \code{"red"}}

\item{max.col.scale}{Rarely requires adjustment. The maximum count of
observations (therefore the darkest shaded color) only used if \code{disp.obs} is
used. This allows consistency of shading between multiple plotted graphs.
It should always be at least as high as the maximum count of observations
plotted}

\item{treat.labs}{A vector of treatment labels in the same order as treatment codes.
Easiest to use treatment labels stored by \code{mb.network()}}

\item{...}{Arguments for \code{ggplot()}}
}
\description{
Plots predicted responses from a time-course MBNMA model
}
\details{
For the S3 method \code{plot()}, if \code{disp.obs} is set to \code{TRUE} it is
advisable to ensure predictions in \code{predict} are estimated using an even
sequence of time points to avoid misrepresentation of shaded densities.
Shaded counts of observations will be relative to the treatment plotted in
each panel rather than to the network reference treatment if \code{disp.obs} is
set to \code{TRUE}.
}
\examples{
\donttest{
# Create an mb.network object from a dataset
copdnet <- mb.network(copd)

# Run an MBNMA model with a log-linear time-course
loglin <- mb.run(copdnet,
  fun=tloglin(pool.rate="rel", method.rate="common"),
  rho="dunif(0,1)", covar="varadj", intercept=FALSE)

# Predict responses using the original dataset to estimate the network reference
#treatment response
df.ref <- copd[copd$treatment=="Placebo",]
predict <- predict(loglin, times=c(0:20), E0=0, ref.resp=df.ref)

# Plot the predicted responses with observations displayed on plot as green shading
plot(predict, disp.obs=TRUE, overlay.ref=FALSE, col="green")

# Plot the predicted responses with the median network reference treatment response overlayed
#on the plot
plot(predict, disp.obs=FALSE, overlay.ref=TRUE)
}

}
