% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.network-class.R
\name{print.mb.network}
\alias{print.mb.network}
\title{Print mb.network information to the console}
\usage{
\method{print}{mb.network}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mb.network}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print mb.network information to the console
}
