% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.predict-class.R
\name{print.mb.predict}
\alias{print.mb.predict}
\title{Print summary information from an mb.predict object}
\usage{
\method{print}{mb.predict}(x, ...)
}
\arguments{
\item{x}{An object of \code{class("mb.predict")} generated by \code{predict.mbnma()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints posterior mean predictions to the console
}
\description{
Print summary information from an mb.predict object
}
