% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.rank-class.R
\name{print.mb.rank}
\alias{print.mb.rank}
\title{Prints a summary of rankings for each parameter}
\usage{
\method{print}{mb.rank}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mb.rank"} generated by \code{rank.mbnma()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints summary details of treatment rankings to the console
}
\description{
Prints a summary of rankings for each parameter
}
