% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sum.functions.R
\name{print.nodesplit}
\alias{print.nodesplit}
\title{Prints basic results from a node-split to the console}
\usage{
\method{print}{nodesplit}(x, groupby = "time.param", ...)
}
\arguments{
\item{x}{An object of class \code{"nodesplit"} generated by \code{mb.nodeplit()}}

\item{groupby}{A character object that can take the value \code{"time.param"} to present
results grouped by time-course parameter (the default) or \code{"comparison"} to present
results grouped by treatment comparison.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Prints basic results from a node-split to the console
}
