% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.array-class.R
\name{print.relative.array}
\alias{print.relative.array}
\title{Print posterior medians (95\% credible intervals) for table of relative effects/mean
differences between treatments/classes}
\usage{
\method{print}{relative.array}(x, digits = 1, ...)
}
\arguments{
\item{...}{further arguments passed to \code{knitr::kable}}
}
\description{
Print posterior medians (95\% credible intervals) for table of relative effects/mean
differences between treatments/classes
}
