% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.functions.R
\name{rank.mb.predict}
\alias{rank.mb.predict}
\title{Rank predictions at a specific time point}
\usage{
\method{rank}{mb.predict}(
  x,
  time = max(x$summary[[1]]$time),
  lower_better = FALSE,
  treats = names(x$summary)
)
}
\arguments{
\item{x}{an object of \code{class("mb.predict")} that contains predictions from an MBNMA model}

\item{time}{a number indicating the time point at which predictions should be ranked. It must
be one of the time points for which predictions in \code{x} are available.}

\item{lower_better}{Indicates whether negative responses are better (\code{lower_better=TRUE}) or}

\item{treats}{A character vector of treatment/class names for which responses have been predicted
in \code{x} As default, rankings will be calculated for all treatments/classes in \code{x}.}
}
\description{
Rank predictions at a specific time point
}
