% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.functions.R
\name{ref.validate}
\alias{ref.validate}
\title{Checks the validity of ref.resp if given as data frame}
\usage{
ref.validate(data.ab)
}
\arguments{
\item{data.ab}{A data frame of arm-level data in "long" format containing the
columns:
\itemize{
\item \code{studyID} Study identifiers
\item \code{time} Numeric data indicating follow-up times
\item \code{y} Numeric data indicating the mean response for a given observation
\item \code{se} Numeric data indicating the standard error for a given observation
}}
}
\description{
Ensures \code{ref.resp} takes the correct form to allow for synthesis of network
reference treatment response if data is provided for meta-analysis
}
