% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.jags.R
\name{remove.loops}
\alias{remove.loops}
\title{Removes any loops from MBNMA model JAGS code that do not contain any
expressions}
\usage{
remove.loops(model)
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}
}
\value{
A character vector of JAGS MBNMA model code that has had empty loops
removed from it
}
\description{
Removes any loops from MBNMA model JAGS code that do not contain any
expressions
}
