% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.network-class.R
\name{summary.mb.network}
\alias{summary.mb.network}
\title{Print summary mb.network information to the console}
\usage{
\method{summary}{mb.network}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mb.network}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints summary details of \code{x} to the console.
}
\description{
Print summary mb.network information to the console
}
