% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sum.functions.R
\name{summary.mb.nodesplit}
\alias{summary.mb.nodesplit}
\title{Takes nodesplit results and produces summary data frame}
\usage{
\method{summary}{mb.nodesplit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"mb.nodesplit"} generated by \code{mb.nodeplit()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
A data frame of summary nodesplit results with the following variables:
\itemize{
\item \code{Comparison} The treatment comparison on which a nodesplit has been performed
\item \code{Time.Param} The time-course parameter on which a nodesplit has been performed
\item \code{Evidence} The evidence contribution for the given comparison (either "Direct" or "Indirect")
\item \code{Median} The posterior median
\item \code{2.5\%} The lower 95% credible interval limit
\item \code{97.5\%} The upper 95% credible interval limit
\item \code{p.value} The Bayesian p-value for the overlap between direct and indirect evidence for
the given comparison (it will therefore have an identical value for direct and indirect evidence
within a particular comparison and time-course parameter)
}
}
\description{
Takes nodesplit results and produces summary data frame
}
