% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma-class.R
\name{summary.mbnma}
\alias{summary.mbnma}
\title{Print summary MBNMA results to the console}
\usage{
\method{summary}{mbnma}(object, ...)
}
\arguments{
\item{object}{An S3 object of \code{class("mbnma")} generated by running
a time-course MBNMA model}

\item{...}{further arguments passed to \code{knitr::kable}}
}
\value{
Prints summary details of the model results to the console
}
\description{
Print summary MBNMA results to the console
}
