% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesplit-class.R
\name{summary.nodesplit}
\alias{summary.nodesplit}
\title{Takes node-split results and produces summary data frame}
\usage{
\method{summary}{nodesplit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"nodesplit"} generated by \code{mb.nodeplit()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
A data frame of summary node-split results with the following variables:
\itemize{
\item \code{Comparison} The treatment comparison on which a node-split has been performed
\item \code{Time.Param} The time-course parameter on which a node-split has been performed
\item \code{Evidence} The evidence contribution for the given comparison (either "Direct" or "Indirect")
\item \code{Median} The posterior median
\item \verb{2.5\%} The lower 95\% credible interval limit
\item \verb{97.5\%} The upper 95\% credible interval limit
\item \code{p.value} The Bayesian p-value for the overlap between direct and indirect evidence for
the given comparison (it will therefore have an identical value for direct and indirect evidence
within a particular comparison and time-course parameter)
}
}
\description{
Takes node-split results and produces summary data frame
}
