% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{time.fun}
\alias{time.fun}
\title{Write time-course function component of JAGS code for MBNMA time-course
models}
\usage{
\method{time}{fun}(
  fun = "linear",
  user.fun = NULL,
  alpha = "arm",
  knots = 3,
  beta.1 = "rel.common",
  beta.2 = NULL,
  beta.3 = NULL,
  beta.4 = NULL
)
}
\arguments{
\item{fun}{An object of class \code{"timefun"} generated (see Details) using any of
\code{tloglin()}, \code{tpoly()}, \code{texp()}, \code{temax()}, \code{tfpoly()}, \code{tspline()} or \code{tuser()}}

\item{alpha}{Refers to the baseline mean response and is a character object
that can take either:
\itemize{
\item \code{"study"} to constrain baseline to be equal for all
arms within a study - reasonable assumption under randomisation
\item \code{"arm"} to allow baseline to vary between arms within a study
}}
}
\value{
A single character string containing JAGS model representing the
time-course function component of an MBNMA time-course model, generated
based on the arguments passed to the function.
}
\description{
Writes a single line of JAGS code representing the time-course function
component of an MBNMA time-course model, outputted as a single character
string.
}
