% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.functions.R
\name{tpoly}
\alias{tpoly}
\title{Polynomial time-course function}
\usage{
tpoly(
  degree = 1,
  pool.1 = "rel",
  method.1 = "common",
  pool.2 = "rel",
  method.2 = "common",
  pool.3 = "rel",
  method.3 = "common",
  pool.4 = "rel",
  method.4 = "common"
)
}
\arguments{
\item{degree}{The degree of the polynomial - e.g. \code{degree=1} for linear, \code{degree=2} for quadratic, \code{degree=3} for cubic.}

\item{pool.1}{Pooling for the 1st polynomial coefficient. Can take \code{"rel"} or \code{"abs"} (see details).}

\item{method.1}{Method for synthesis of the 1st polynomial coefficient. Can take \verb{"common} or \code{"random"} (see details).}

\item{pool.2}{Pooling for the 2nd polynomial coefficient. Can take \code{"rel"} or \code{"abs"} (see details).}

\item{method.2}{Method for synthesis of the 2nd polynomial coefficient. Can take \verb{"common} or \code{"random"} (see details).}

\item{pool.3}{Pooling for the 3rd polynomial coefficient. Can take \code{"rel"} or \code{"abs"} (see details).}

\item{method.3}{Method for synthesis of the 3rd polynomial coefficient. Can take \verb{"common} or \code{"random"} (see details).}

\item{pool.4}{Pooling for the 4th polynomial coefficient. Can take \code{"rel"} or \code{"abs"} (see details).}

\item{method.4}{Method for synthesis of the 4th polynomial coefficient. Can take \verb{"common} or \code{"random"} (see details).}
}
\value{
An object of \code{class("timefun")}
}
\description{
\eqn{\beta_1} represents the 1st coefficient.
\eqn{\beta_2} represents the 2nd coefficient.
\eqn{\beta_3} represents the 3rd coefficient.
\eqn{\beta_4} represents the 4th coefficient.
}
\details{
Linear model:
\deqn{\beta_1{x}}

Quadratic model:
\deqn{\beta_1{x} + \beta_2{x^2}}

Cubic model:
\deqn{\beta_1{x} + \beta_2{x^2} + \beta_3{x^3}}

Quartic model:
\deqn{\beta_1{x} + \beta_2{x^2} + \beta_3{x^3} + \beta_4{x^4}}

Time-course parameters in the model must be specified using a \code{pool} and a \code{method} prefix.

\code{pool} is used to define the approach used for pooling of a given time-course parameter and
can take either:
\itemize{
\item \code{"rel"} indicates that relative effects should be pooled for this time-course parameter.
This preserves randomisation within included studies, are likely to vary less between studies (only due to effect modification),
and allow for testing of consistency between direct and indirect evidence (e.g. using \code{mb.nodesplit()}). Pooling follows the
general approach for Network Meta-Analysis proposed by \insertCite{lu2004;textual}{MBNMAtime}.
\item \code{"abs"} indicates that study arms should be pooled across the whole network for this
time-course parameter  \emph{independently of assigned treatment}.
This implies estimating a single absolute value across the network for this time-course parameter,
and may therefore be making strong assumptions of similarity.
}

\code{method} is used to define the model used for meta-analysis for a given time-course parameter
and can take any of the following values:
\itemize{
\item \code{"common"} implies that all studies estimate the same true effect
(akin to a "fixed effects" meta-analysis)
\item \code{"random"} implies that all studies estimate a separate true effect, but that each
of these true effects vary randomly around a true mean effect. This approach allows
for modelling of between-study heterogeneity.
}

When relative effects are modelled on more than one time-course parameter,
correlation between the time-course parameters is automatically
estimated using a vague Wishart prior. This prior can be made slightly more informative
by specifying the relative scale of variances between the time-course parameters in \code{mb.run()} using
\code{var.scale}.
}
\examples{
# Linear model with random effects
tpoly(pool.1="rel", method.1="random")

# Quadratic model with a single absolute parameter estimated for the 2nd coefficient
tpoly(pool.1="rel", method.1="common", pool.2="abs", method.2="random")

}
\references{
\insertAllCited
}
