% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.jags.R
\name{write.check}
\alias{write.check}
\title{Checks validity of arguments for mb.write}
\usage{
write.check(
  fun = linear(),
  positive.scale = TRUE,
  intercept = TRUE,
  rho = NULL,
  covar = NULL,
  var.scale = NULL,
  link = "identity",
  class.effect = list(),
  UME = c()
)
}
\arguments{
\item{fun}{is a character specifying a functional form to be assigned to the
time-course. Options are given in \code{details}.}

\item{positive.scale}{A boolean object that indicates whether all continuous
mean responses (y) are positive and therefore whether the baseline response
should be given a prior that constrains it to be positive.}

\item{intercept}{A boolean object that indicates whether an intercept is to
be included in the model. Can be used to imply whether mean responses in
data are change from baseline (\code{FALSE}) or not (setting it to \code{FALSE}
removes the intercept, \code{alpha}, from the model).}

\item{rho}{The correlation coefficient when modelling correlation between time points. The default is a string representing a
prior distribution indicating that it be estimated from the data (default: \code{rho="dunif(0,1)"}). If left
as \code{NULL} then this implies modelling no correlation between time points. Can also be assigned a numeric value
(e.g. \code{rho=0.7}), which fixes \code{rho} in the model to this value (e.g. for use in a deterministic sensitivity analysis).}

\item{covar}{A character specifying the covariance structure to use for modelling correlation between time-points. This can
be done by specifying one of the following:
\itemize{
\item \code{"varadj"} - a univariate likelihood with a variance adjustment to assume a constant correlation between subsequent
time points \insertCite{jansen2015}{MBNMAtime}. This is the default.
\item \code{"CS"} - a multivariate normal likelihood with a compound symmetry structure
\item \code{"AR1"} - a multivariate normal likelihood with an autogregressive AR1 structure
}}

\item{var.scale}{A numeric vector indicating the relative scale of variances between
correlated time-course parameters when relative effects are modelled on more than
one time-course parameter(see Details LINK). Each element of
the vector refers to the relative scale of each of the time-course parameters that is
modelled using relative effects.}

\item{link}{Can take either \code{"identity"}, \code{"log"} (for modelling Ratios of Means \insertCite{friedrichROM}{MBNMAtime}) or
\code{"smd"} (for modelling Standardised Mean Differences).}

\item{class.effect}{A list of named strings that determines which time-course
parameters to model with a class effect and what that effect should be
(\code{"common"} or \code{"random"}). For example: \code{list("beta.2"="common", "beta.3"="random")}.}

\item{UME}{Can take either \code{TRUE} or \code{FALSE} (for an unrelated mean effects
model on all or no time-course parameters respectively) or can be a vector
of parameter name strings to model as UME. For example: \code{c("beta.1", "beta.2")}.}
}
\value{
A boolean object that indicates whether the arguments imply modelling
correlation between time points.
}
\description{
Checks validity of arguments for mb.write
}
\details{
Used to check if the arguments given to mb.write are valid. The
function will return informative errors if arguments are mispecified and
will return an object that indicates whether the arguments imply modelling a
correlation between time points if it passes.
}
