% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.fract.poly}
\alias{write.fract.poly}
\title{Adds additional sections of JAGS code for an MBNMA model required for models
with a fractional polynomial time-course}
\usage{
write.fract.poly(model, timecourse)
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}

\item{timecourse}{A character object representing the time-course used in the MBNMA model}
}
\value{
A character object of JAGS MBNMA model code that includes fractional
polynomial components of the model
}
\description{
Adds additional sections of JAGS code for an MBNMA model required for models
with a fractional polynomial time-course
}
\details{
FUNCTION IS DEPRECATED
}
