% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.jags.R
\name{write.model}
\alias{write.model}
\title{Write the basic JAGS model code for MBNMA to which other lines of model
code can be added}
\usage{
write.model()
}
\value{
A character vector of JAGS model code
}
\description{
Write the basic JAGS model code for MBNMA to which other lines of model
code can be added
}
