% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.piecelinear}
\alias{write.piecelinear}
\title{Adds additional sections of JAGS code for an MBNMA model required for models
with a piecewise linear time-course}
\usage{
write.piecelinear(model, beta.1, beta.3)
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}

\item{beta.1}{A list with named elements \code{pool} and \code{method} that refers to
time-course parameter(s) specified within the time-course function (see details).}

\item{beta.3}{A list with named elements \code{pool} and \code{method} that refers to
time-course parameter(s) specified within the time-course function (see details).}
}
\value{
A character object of JAGS MBNMA model code that includes piecewise
linear components of the model
}
\description{
Adds additional sections of JAGS code for an MBNMA model required for models
with a piecewise linear time-course
}
\details{
FUNCTION DEPRACATED
}
