% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.jags.R
\name{write.rw}
\alias{write.rw}
\title{Write non-parametric random walk model JAGS code}
\usage{
write.rw(method = "common", link = "identity")
}
\arguments{
\item{link}{Can take either \code{"identity"} (the default),
\code{"log"} (for modelling Ratios of Means \insertCite{friedrich2011}{MBNMAtime}) or
\code{"smd"} (for modelling Standardised Mean Differences - although this also corresponds to an identity link function).}
}
\value{
A single long character string containing the JAGS model generated
based on the arguments passed to the function.
}
\description{
Writes JAGS code for a Bayesian non-parametric model that splits the
data into different time-bins and assumes a random walk process for treatment
effects between them.
}
\examples{
# Write a common effects non-paramtric random walk model
write.rw(method="common")

}
