% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot.timefun}
\alias{plot.timefun}
\title{Plot illustrative time-course functions}
\usage{
\method{plot}{timefun}(x = tpoly(degree = 1), beta.1 = 0, beta.2 = 0, beta.3 = 0, beta.4 = 0)
}
\arguments{
\item{x}{An object of \code{class("timefun")}}
}
\value{
An illustrative plot of the time-course function with the parameters specified
in \code{beta.1}, \code{beta.2}, \code{beta.3} and \code{beta.4}
}
\description{
Can be used to plot potential time-course functions to identify if they may
be a suitable fit for the data.
}
